backend/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
SQLAlchemy==2.0.21
alembic==1.12.0
asyncpg==0.28.0
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
chromadb==0.4.14
redis==5.0.1
python-multipart==0.0.6
PyMuPDF==1.23.6
sentence-transformers==2.2.2
langchain==0.0.330
anthropic==0.5.0
aiofiles==23.2.1
tiktoken==0.5.1
pytest==7.4.3
pytest-asyncio==0.21.1
redis==5.0.1
